+++ src/main/kotlin/com/github/jpvos/keylogger/KeyloggerBundle.kt
@@ -5,10 +5,10 @@ import org.jetbrains.annotations.NonNls
 import org.jetbrains.annotations.PropertyKey
 
 @NonNls
-private const val BUNDLE = "messages.MyBundle"
+private const val BUNDLE = "messages.KeyloggerBundle"
 
-object MyBundle : DynamicBundle(BUNDLE) {
 
+object KeyloggerBundle : DynamicBundle(BUNDLE) {
     @JvmStatic
     fun message(@PropertyKey(resourceBundle = BUNDLE) key: String, vararg params: Any) =
         getMessage(key, *params)