+++ src/main/kotlin/com/github/jpvos/keylogger/event/ActionEvent.kt
@@ -1,0 +1,13 @@
+package com.github.jpvos.keylogger.event
+
+
+data class ActionEvent(
+    val type: Type,
+    val name: String
+) {
+    enum class Type {
+        Mouse,
+        Keystroke,
+        EditorAction
+    }
+}