+++ src/main/kotlin/com/github/jpvos/keylogger/event/CounterEvent.kt
@@ -1,0 +1,7 @@
+package com.github.jpvos.keylogger.event
+
+data class CounterEvent(
+    val actions: Map<ActionEvent, Long>,
+    val totalActiveTime: Long,
+    val totalIdleTime: Long
+)