+++ src/main/kotlin/com/github/jpvos/keylogger/util/Listener.kt
@@ -1,0 +1,6 @@
+package com.github.jpvos.keylogger.util
+
+// TODO: Replace with some built-in idiomatic listener class
+interface Listener<Event> {
+    fun onEvent(event: Event)
+}