+++ CHANGELOG.md
@@ -5,3 +5,9 @@
 ## [Unreleased]
 ### Added
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+
+## [Unreleased]
+### Added
+- Prototype implementation of action/click/type listeners
+- Simple tool window to display actions per minute
+- README and CHANGELOG